Adds the backend of the app
The app consists of a basic api that have CRUD operations on the domnain Phonebook. The app counts with a clean architecture and has some custom validations inside the entrypoint level. The app also counts with PMD for error checking and spotless to automatically format the code based on the format.